Fixup tests
Pull out some utilities that are useful across multiple tests, and
switch to using @testing-library/react, which is recommended by the
React documentation.
