+++ package.json
@@ -58,6 +58,7 @@
     "use-persisted-state": "^0.3.3"
   },
   "devDependencies": {
+    "@testing-library/react": "^12.1.2",
     "@types/chai": "^4.3.0",
     "@types/dompurify": "^2.3.3",
     "@types/element-resize-detector": "^1.1.3",
